+++ mayan/apps/templating/widgets.py
@@ -17,12 +17,6 @@ class TemplateWidget(NamedMultiWidget):
     def __init__(self, attrs=None, **kwargs):
         widgets = OrderedDict()
 
-        widgets['model_attribute'] = forms.widgets.Select(
-            attrs={
-                'data-autocopy': 'true',
-                'data-field-template': '{{ ${ $idTemplate.data("model-variable") }.${ $this.val() } }}'
-            }
-        )
         widgets['builtin_tags'] = forms.widgets.Select(
             attrs={
                 'data-autocopy': 'true',
@@ -32,9 +26,7 @@ class TemplateWidget(NamedMultiWidget):
         widgets['template'] = forms.widgets.Textarea(
             attrs={'rows': 5, 'data-template-fields': 'template'}
         )
-        super(TemplateWidget, self).__init__(
-            widgets=widgets, attrs=attrs, **kwargs
-        )
+        super().__init__(widgets=widgets, attrs=attrs, **kwargs)
 
     def get_builtin_choices(self, klass, name_template='{}'):
         result = []
@@ -59,7 +51,91 @@ class TemplateWidget(NamedMultiWidget):
         return result
 
     def get_context(self, name, value, attrs):
-        result = super(TemplateWidget, self).get_context(name, value, attrs)
+        result = super().get_context(name, value, attrs)
+        # Set each autocopy sub widget as not required
+        result['widget']['subwidgets'][0]['attrs']['required'] = False
+        #result['widget']['subwidgets'][1]['attrs']['required'] = False
+        return result
+
+    def decompress(self, value):
+        choices_builtin = []
+        choices_builtin.append(
+            (
+                _('Filters'), self.get_builtin_choices(
+                    klass='filters', name_template='{{{{ | {} }}}}'
+                )
+            )
+        )
+        choices_builtin.append(
+            (
+                _('Tags'), self.get_builtin_choices(
+                    klass='tags', name_template='{{% {} %}}'
+                )
+            )
+        )
+        choices_builtin.insert(
+            0, ('', _('<Filters and tags>'))
+        )
+
+        self.widgets['builtin_tags'].choices = choices_builtin
+        return {
+            'builtin_tags': None, 'template': value
+        }
+
+    def value_from_datadict(self, querydict, files, name):
+        template = querydict.get('{}_template'.format(name))
+
+        return template
+
+
+class ModelTemplateWidget(TemplateWidget):
+    class Media:
+        js = ('templating/js/template_widget.js',)
+
+    def __init__(self, attrs=None, **kwargs):
+        widgets = OrderedDict()
+
+        widgets['model_attribute'] = forms.widgets.Select(
+            attrs={
+                'data-autocopy': 'true',
+                'data-field-template': '{{ ${ $idTemplate.data("model-variable") }.${ $this.val() } }}'
+            }
+        )
+        widgets['builtin_tags'] = forms.widgets.Select(
+            attrs={
+                'data-autocopy': 'true',
+                'data-field-template': '${ $this.val() }',
+            }
+        )
+        widgets['template'] = forms.widgets.Textarea(
+            attrs={'rows': 5, 'data-template-fields': 'template'}
+        )
+        super().__init__(widgets=widgets, attrs=attrs, **kwargs)
+
+    def get_builtin_choices(self, klass, name_template='{}'):
+        result = []
+        template = Template('')
+        builtin_libraries = [
+            ('', library) for library in template._template.engine.template_builtins
+        ]
+        for module_name, library in builtin_libraries:
+            for name, function in getattr(library, klass).items():
+                title, body, metadata = admindocs.utils.parse_docstring(
+                    function.__doc__
+                )
+                title = _(title)
+                result.append(
+                    (
+                        name_template.format(name), '{} - {}'.format(name, title)
+                    )
+                )
+
+        result = sorted(result, key=lambda x: x[0])
+
+        return result
+
+    def get_context(self, name, value, attrs):
+        result = super().get_context(name, value, attrs)
         # Set each autocopy sub widget as not required
         result['widget']['subwidgets'][0]['attrs']['required'] = False
         result['widget']['subwidgets'][1]['attrs']['required'] = False
@@ -69,10 +145,7 @@ class TemplateWidget(NamedMultiWidget):
         attribute_choices = ModelAttribute.get_all_choices_for(
             model=self.attrs['model']
         )
-        choices = []
-
-        choices = attribute_choices
-        choices.insert(
+        attribute_choices.insert(
             0, ('', _('<Model attributes>'))
         )
 
@@ -96,7 +169,7 @@ class TemplateWidget(NamedMultiWidget):
         )
 
         self.widgets['builtin_tags'].choices = choices_builtin
-        self.widgets['model_attribute'].choices = choices
+        self.widgets['model_attribute'].choices = attribute_choices
         return {
             'builtin_tags': None, 'model_attribute': None,
             'template': value