+++ mayan/apps/web_links/forms.py
@@ -1,7 +1,8 @@
 from django import forms
+from django.utils.text import format_lazy
 
 from mayan.apps.documents.models import Document
-from mayan.apps.templating.fields import TemplateField
+from mayan.apps.templating.fields import ModelTemplateField
 
 from .models import WebLink
 
@@ -9,9 +10,10 @@ from .models import WebLink
 class WebLinkForm(forms.ModelForm):
     def __init__(self, *args, **kwargs):
         super(WebLinkForm, self).__init__(*args, **kwargs)
-        self.fields['template'] = TemplateField(
-            initial_help_text=self.fields['template'].help_text,
-            label=self.fields['template'].label, model=Document,
+        self.fields['template'] = ModelTemplateField(
+            initial_help_text=format_lazy(
+                '{} ', self.fields['template'].help_text
+            ), label=self.fields['template'].label, model=Document,
             model_variable='document', required=True
         )